requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
aniso8601==8.0.0
astor==0.8.1
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
discord.py==1.3.4
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
nltk==3.5
numpy==1.19.1
pdfcrowd==4.12.0
pdfkit==0.6.1
Pillow==7.2.0
protobuf==3.12.4
pycparser==2.20
pylint==2.5.3
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tflearn==0.3.2
toml==0.10.1
tqdm==4.48.2
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
yarl==1.5.1
zipp==3.1.0
